Fix: correct Sass import of bootstap-datepicker styles
This was a mistake in the monorepo setup. The `.css` extension in an import causes the file to be excluded from compilation. The remainder of this commit:

- shares path resolution for Sass imports between Core and Express
- ensures the bootstrap-datepicker dependency isn’t hoisted to the monorepo root, for a predictable path in the Sass file
- adds a current snapshot of the all-widgets ODK form fixture, to conveniently verify the fix
- autoformatting of `.eslintrc.json`
