+++ packages/enketo-core/src/widget/date/datepicker-extended.scss
@@ -1,4 +1,4 @@
-@import '/node_modules/bootstrap-datepicker/dist/css/bootstrap-datepicker.min.css';
+@import '/packages/enketo-core/node_modules/bootstrap-datepicker/dist/css/bootstrap-datepicker.min';
 
 /** fixes by martijn **/
